# openapi.yaml
openapi: 3.0.3
info:
  title: E-commerce Lite API
  description: RESTful API for a simple e-commerce application (course project)
  version: 1.0.0
servers:
  - url: http://localhost:5000/api/v1
    description: Local development server

paths:
  /auth/register:
    post:
      summary: Register a new user account
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 80
                password:
                  type: string
                  minLength: 6
      responses:
        '201':
          description: User created, returns JWT token
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  token:
                    type: string
        '400':
          description: Validation error
        '409':
          description: Username already taken

  /auth/login:
    post:
      summary: Login with username and password
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful, returns JWT token
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  token:
                    type: string
        '401':
          description: Invalid credentials
        '400':
          description: Invalid request body

  /products:
    get:
      summary: List all products (paginated)
      tags: [Products]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Product list
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  total:
                    type: integer
                  pages:
                    type: integer
                  current_page:
                    type: integer
    post:
      summary: Create a new product (admin only)
      tags: [Products]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, price]
              properties:
                name:
                  type: string
                description:
                  type: string
                price:
                  type: number
                  minimum: 0
                stock:
                  type: integer
                  minimum: 0
                  default: 0
                category_id:
                  type: integer
      responses:
        '201':
          description: Product created
        '400':
          description: Validation error
        '401':
          description: Not authenticated
        '403':
          description: Admin access required

  /products/{id}:
    get:
      summary: Get a single product by ID
      tags: [Products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Product details
        '404':
          description: Product not found
    put:
      summary: Update a product (admin only)
      tags: [Products]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                price:
                  type: number
                stock:
                  type: integer
                category_id:
                  type: integer
      responses:
        '200':
          description: Product updated
        '404':
          description: Product not found
    delete:
      summary: Delete a product (admin only)
      tags: [Products]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Product deleted
        '401':
          description: Not authenticated

  /categories:
    get:
      summary: List all categories
      tags: [Categories]
      responses:
        '200':
          description: Category list
    post:
      summary: Create a category (admin only)
      tags: [Categories]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Category created
        '409':
          description: Category already exists

  /categories/{id}:
    put:
      summary: Update a category (admin only)
      tags: [Categories]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Category updated
    delete:
      summary: Delete a category (admin only)
      tags: [Categories]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Category deleted

  /orders:
    get:
      summary: List all orders (admin only)
      tags: [Orders]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Order list
        '403':
          description: Admin access required
    post:
      summary: Create a new order
      tags: [Orders]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  items:
                    type: object
                    required: [product_id, quantity]
                    properties:
                      product_id:
                        type: integer
                      quantity:
                        type: integer
      responses:
        '201':
          description: Order created
        '400':
          description: Validation or stock error

  /orders/mine:
    get:
      summary: List current user's orders
      tags: [Orders]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User's order list

  /orders/{id}/simulate-payment:
    post:
      summary: Mark an order as paid (customer only)
      tags: [Orders]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Order paid
        '403':
          description: Not your order
        '400':
          description: Already paid or invalid status

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        stock:
          type: integer
        category_id:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time